% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_file_ops.R
\docType{data}
\name{s3_file_ops}
\alias{s3_file_ops}
\alias{R6_s3_file_ops}
\title{S3 file op equivalents}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
s3_file_ops(bucket)

R6_s3_file_ops
}
\arguments{
\item{bucket}{Name of the S3 bucket for which you wish to create or connect
store}
}
\keyword{datasets}
