% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{storr_s3}
\alias{storr_s3}
\alias{driver_s3}
\title{S3 backend for rds object cache driver}
\usage{
storr_s3(bucket, remote_root, ..., path_local = NULL,
  default_namespace = "objects")

driver_s3(bucket, remote_root, ..., path_local = NULL)
}
\arguments{
\item{bucket}{Name of the S3 bucket for which you wish to create
of connect store}

\item{remote_root}{Base path on the remote bucket}

\item{...}{Additional arguments passed through to
\code{storr::driver_rds}}

\item{path_local}{Optional path to a local cache (see
\code{storr::driver_remote})}

\item{default_namespace}{Default namespace (see \code{storr::storr})}
}
