% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{storr_ssh}
\alias{storr_ssh}
\alias{driver_ssh}
\title{ssh storr}
\usage{
storr_ssh(session, remote_root, ..., path_local = NULL,
  default_namespace = "default_namespace")

driver_ssh(session, remote_root, ..., path_local = NULL)
}
\arguments{
\item{session}{A ssh session object (see \code{ssh::ssh_connect})}

\item{remote_root}{Base path on the remote bucket}

\item{...}{Additional arguments passed through to
\code{storr::driver_rds}}

\item{path_local}{Optional path to a local cache (see
\code{storr::driver_remote})}

\item{default_namespace}{Default namespace (see \code{storr::storr})}
}
\description{
ssh backend for storr
}
